# coprogo/src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: coprogo
  version: 1.0.0
paths:
...
  /api/groups/{groupId}/expenses:
    get:
      tags:
      - Expenses
      summary: List expenses for a group
      operationId: listGroupExpenses
      parameters:
      - name: groupId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: listGroupExpenses 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpenseResponse"
      security:
      - bearerAuth: []
    post:
      tags:
      - Expenses
      summary: Propose an expense
      operationId: proposeExpense
      parameters:
      - name: groupId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposeExpenseRequest"
        required: true
      responses:
        "409":
          description: Supporting document upload intent is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "204":
          description: proposeExpense 204 response
      security:
      - bearerAuth: []
  /api/groups/{groupId}/expenses/{expenseId}:
    get:
      tags:
      - Expenses
      summary: Get expense details
      operationId: getExpenseDetail
      parameters:
      - name: groupId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: expenseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: getExpenseDetail 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseDetailResponse"
      security:
      - bearerAuth: []
  /api/groups/{groupId}/expenses/{expenseId}/participation-decisions:
    post:
      tags:
      - Expenses
      summary: Record a participation decision for an expense
      operationId: recordParticipationDecision
      parameters:
      - name: groupId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: expenseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseParticipationDecisionRequest"
        required: true
      responses:
        "204":
          description: recordParticipationDecision 204 response
      security:
      - bearerAuth: []
...
  /api/groups/{groupId}/expenses/{expenseId}/supporting-documents:
    get:
      tags:
      - Supporting documents
      summary: List current and historical supporting documents for an expense
      operationId: listExpenseSupportingDocuments
      parameters:
      - name: groupId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: expenseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Supporting documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseSupportingDocumentsResponse"
      security:
      - bearerAuth: []
...
  /api/groups/{groupId}/supporting-document-uploads:
    post:
      tags:
      - Supporting document uploads
      summary: Start a supporting document upload
      operationId: startUpload
      parameters:
      - name: groupId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportingDocumentUploadRequest"
        required: true
      responses:
        "201":
          description: Supporting document upload target created
          headers:
            Location:
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportingDocumentUploadResponse"
      security:
      - bearerAuth: []
  /api/groups/{groupId}/supporting-document-uploads/{uploadIntentId}/confirmation:
    post:
      tags:
      - Supporting document uploads
      summary: Confirm a supporting document upload
      operationId: confirmUpload
      parameters:
      - name: groupId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: uploadIntentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: confirmUpload 204 response
      security:
      - bearerAuth: []
...
components:
  schemas:
    ApiErrorResponse:
      required:
      - message
      - path
      type: object
      properties:
        message:
          type: string
        path:
          type: string
...
    CumulativeExpenseTierRequest:
      required:
      - participants
      - upToAmountInCents
      type: object
      properties:
        upToAmountInCents:
          minimum: 0
          exclusiveMinimum: true
          type: integer
          format: int64
        participants:
          minItems: 1
          type: array
          items:
            minLength: 1
            type: string
            format: email
    CumulativeTiersExpenseAllocationRequest:
      required:
      - tiers
      - type
      type: object
      properties:
        tiers:
          minItems: 1
          type: array
          items:
            $ref: "#/components/schemas/CumulativeExpenseTierRequest"
        type:
          type: string
          enum:
          - CUMULATIVE_TIERS
    CustomExpenseAllocationRequest:
      required:
      - participations
      - type
      type: object
      properties:
        participations:
          minItems: 1
          type: array
          items:
            $ref: "#/components/schemas/CustomExpenseParticipationRequest"
        type:
          type: string
          enum:
          - CUSTOM
    CustomExpenseParticipationRequest:
      required:
      - amountInCents
      - member
      type: object
      properties:
        member:
          minLength: 1
          type: string
          format: email
        amountInCents:
          minimum: 0
          exclusiveMinimum: true
          type: integer
          format: int64
...
    EqualExpenseAllocationRequest:
      required:
      - participants
      - type
      type: object
      properties:
        participants:
          minItems: 1
          type: array
          items:
            minLength: 1
            type: string
            format: email
        type:
          type: string
          enum:
          - EQUAL
    EqualWithCapsExpenseAllocationRequest:
      required:
      - caps
      - participants
      - type
      type: object
      properties:
        participants:
          minItems: 1
          type: array
          items:
            minLength: 1
            type: string
            format: email
        caps:
          minItems: 1
          type: array
          items:
            $ref: "#/components/schemas/ExpenseAllocationCapRequest"
        type:
          type: string
          enum:
          - EQUAL_WITH_CAPS
    ExpenseAllocationCapRequest:
      required:
      - maximumAmountInCents
      - member
      type: object
      properties:
        member:
          minLength: 1
          type: string
          format: email
        maximumAmountInCents:
          minimum: 0
          exclusiveMinimum: true
          type: integer
          format: int64
    ExpenseAllocationRequest:
      type: object
      properties:
        type:
          type: string
      discriminator:
        propertyName: type
        mapping:
          EQUAL_WITH_CAPS: "#/components/schemas/EqualWithCapsExpenseAllocationRequest"
          EQUAL: "#/components/schemas/EqualExpenseAllocationRequest"
          CUSTOM: "#/components/schemas/CustomExpenseAllocationRequest"
          CUMULATIVE_TIERS: "#/components/schemas/CumulativeTiersExpenseAllocationRequest"
      oneOf:
      - $ref: "#/components/schemas/EqualExpenseAllocationRequest"
      - $ref: "#/components/schemas/EqualWithCapsExpenseAllocationRequest"
      - $ref: "#/components/schemas/CumulativeTiersExpenseAllocationRequest"
      - $ref: "#/components/schemas/CustomExpenseAllocationRequest"
    ExpenseDetailParticipationResponse:
      required:
      - amountCents
      - member
      - status
      type: object
      properties:
        member:
          type: string
        amountCents:
          type: integer
          format: int64
        status:
          type: string
    ExpenseDetailResponse:
      required:
      - createdAt
      - createdBy
      - id
      - participations
      - status
      - title
      - totalAmountCents
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        createdBy:
          type: string
        totalAmountCents:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        status:
          type: string
        participations:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseDetailParticipationResponse"
        refusal:
          nullable: true
          allOf:
          - $ref: "#/components/schemas/ExpenseRefusalResponse"
    ExpenseParticipationDecisionInput:
      type: string
      enum:
      - APPROVE
      - REFUSE
    ExpenseParticipationDecisionRequest:
      required:
      - decision
      type: object
      properties:
        decision:
          $ref: "#/components/schemas/ExpenseParticipationDecisionInput"
        reason:
          type: string
          nullable: true
    ExpenseRefusalResponse:
      required:
      - member
      - refusedAt
      type: object
      properties:
        member:
          type: string
        refusedAt:
          type: string
          format: date-time
        reason:
          type: string
          nullable: true
    ExpenseResponse:
      required:
      - createdAt
      - createdBy
      - id
      - status
      - title
      - totalAmountCents
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        createdBy:
          type: string
        totalAmountCents:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        status:
          type: string
        refusal:
          nullable: true
          allOf:
          - $ref: "#/components/schemas/ExpenseRefusalResponse"
...
    ExpenseSupportingDocumentResponse:
      required:
      - attachedAt
      - canDelete
      - download
      - fileName
      - mediaType
      - sizeBytes
      - sourceUploadIntent
      - uploader
      type: object
      properties:
        sourceUploadIntent:
          type: string
          format: uuid
        fileName:
          type: string
        mediaType:
          type: string
        sizeBytes:
          type: integer
          format: int64
        uploader:
          type: string
        attachedAt:
          type: string
          format: date-time
        replacesSourceUploadIntent:
          type: string
          format: uuid
          nullable: true
        deletion:
          nullable: true
          allOf:
          - $ref: "#/components/schemas/SupportingDocumentDeletionResponse"
        canDelete:
          type: boolean
        download:
          $ref: "#/components/schemas/SupportingDocumentDownloadResponse"
    ExpenseSupportingDocumentsResponse:
      required:
      - current
      - history
      type: object
      properties:
        current:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseSupportingDocumentResponse"
        history:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseSupportingDocumentResponse"
...
    ProposeExpenseRequest:
      required:
      - allocation
      - title
      - totalAmountInCents
      type: object
      properties:
        title:
          minLength: 1
          type: string
        totalAmountInCents:
          minimum: 0
          exclusiveMinimum: true
          type: integer
          format: int64
        allocation:
          $ref: "#/components/schemas/ExpenseAllocationRequest"
        supportingDocumentUploadIntents:
          type: array
          items:
            type: string
            format: uuid
...
    SupportingDocumentDeletionResponse:
      required:
      - deletedAt
      - deletedBy
      type: object
      properties:
        deletedBy:
          type: string
        deletedAt:
          type: string
          format: date-time
    SupportingDocumentDownloadResponse:
      required:
      - expiresAt
      - url
      type: object
      properties:
        url:
          type: string
          format: uri
        expiresAt:
          type: string
          format: date-time
    SupportingDocumentUploadRequest:
      required:
      - fileName
      - mediaType
      - sha256
      - sizeBytes
      type: object
      properties:
        fileName:
          minLength: 1
          type: string
        mediaType:
          minLength: 1
          type: string
        sizeBytes:
          minimum: 0
          exclusiveMinimum: true
          type: integer
          format: int64
        sha256:
          minLength: 1
          type: string
    SupportingDocumentUploadResponse:
      required:
      - expiresAt
      - intentId
      - requiredHeaders
      - uploadUrl
      type: object
      properties:
        intentId:
          type: string
          format: uuid
        uploadUrl:
          type: string
          format: uri
        requiredHeaders:
          type: object
          additionalProperties:
            type: string
        expiresAt:
          type: string
          format: date-time